Split up evaluation_kernels.h into cell and face parts
