+++ include/deal.II/matrix_free/evaluation_template_factory_internal.h
@@ -28,6 +28,21 @@ DEAL_II_NAMESPACE_OPEN
 
 namespace internal
 {
+  /**
+   * This struct is used to implement
+   * FEEvaluation::fast_evaluation_supported() and
+   * FEFaceEvaluation::fast_evaluation_supported().
+   */
+  struct FastEvaluationSupported
+  {
+    template <int fe_degree, int n_q_points_1d>
+    static bool
+    run()
+    {
+      return fe_degree != -1;
+    }
+  };
+
   template <int degree, typename EvaluatorType, typename... Args>
   bool
   instantiation_helper_run(const unsigned int given_degree,
